requirements.txt
streamlit==1.51.0
python-dotenv
langchain==1.0.3
langchain-groq==1.0.0
langchain-community==0.4.1
langchain-huggingface==1.0.0
qdrant-client==1.15.1
PyGithub>=1.59.0
sentence-transformers==5.1.2
fastembed==0.7.3
